Add raw mode support to JS and wasm
