+++ buildSrc/src/main/kotlin/mordant-js-sample-conventions.gradle.kts
@@ -1,0 +1,17 @@
+import org.jetbrains.kotlin.gradle.targets.js.dsl.ExperimentalWasmDsl
+
+plugins {
+    kotlin("multiplatform")
+}
+
+kotlin {
+    js  {
+        nodejs()
+        binaries.executable()
+    }
+    @OptIn(ExperimentalWasmDsl::class)
+    wasmJs {
+        nodejs()
+        binaries.executable()
+    }
+}
