+++ buildSrc/src/main/kotlin/mordant-mpp-sample-conventions.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     id("mordant-jvm-sample-conventions")
     id("mordant-native-sample-conventions")
+    id("mordant-js-sample-conventions")
 }
 
 kotlin {
